ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pictobrickwebapp"  # you can rename later to ptb-ml if you want
version = "0.1.0"
requires-python = ">=3.11, <3.13"
dependencies = [
    "numpy<2",
    "pandas>=2.3.3",
    # preprocessing engine deps (Milestone 1)
    "pillow>=10",
    "opencv-python-headless>=4.8",
    "imagehash>=4.3",
    "pydantic>=2,<3",
    "pillow-heif>=1.2.1",
    "pytest>=9.0.2",
    "torch>=2.0, <=2.2.2",
    "torchvision>=0.15, <=0.17.2",
    "transformers>=4.40,<5.0",
    "accelerate>=0.30",
    "geffnet>=1.0.2",
    "open3d>=0.19.0",
    "pygltflib>=1.16.5",
    "sqlalchemy>=2.0.49",
    # API + worker deps (Phase 1)
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "celery[redis]>=5.3",
    "boto3>=1.34",
    "psycopg2-binary>=2.9",
    # Auth deps (Phase 2)
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src", "vendor"]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"
name = "pytorch-cpu"
explicit = true

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cpu", marker = "sys_platform == 'darwin' and platform_machine == 'arm64'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cpu", marker = "sys_platform == 'darwin' and platform_machine == 'arm64'" },
]
